[OPAL-437] Have feedback on product add
On adding a Hub product to a tenant, the product selection
dialog closed immediately and there was no feedback provided
to the user for 8+ seconds.

This change provides immediate feedback to the user by
clearing the existing products list and showing the loading
indicator. An interim toastr message confirms the products
are added.

A slight tweak to the loading of products within the selection
dialog box is also added for consistency.