Remove user from list-org list
On removing a user from the organisation's list of members, update the
the source list, and then refilter the display list.

Refactored the search code for conciseness, and so it could be called
from the removal code.

Also tweaked the name displayed on successfully adding a user. If the
user already exists, the name will the be the existing record's user
name, not the one the user enters.